requirements.txt
black
jinja2
mkdocs-get-deps
mkdocs-material
pyyaml
shot-scraper
tqdm
